app/bot.py
import os
import asyncio

from telegram import Update
from telegram.ext import ApplicationBuilder, ContextTypes, CommandHandler


class HomeserverBot:

    def __init__(self):
        self.application = ApplicationBuilder().token(os.getenv("TELEGRAM_BOT_TOKEN")).build()

        start_handler = CommandHandler('start', self.start)
        self.application.add_handler(start_handler)

    def run(self):
        self.application.run_polling()

    async def start(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
        await context.bot.send_message(
            chat_id=update.effective_chat.id,
            text="I'm a bot, please talk to me!"
        )

if __name__ == '__main__':
    bot = HomeserverBot()
    bot.run()
